+++ src/app/detailed-targeting/detailed-targeting-api/detailed-targeting-api.service.ts
@@ -5,6 +5,8 @@ import { DetailedTargetingInfoService } from '../detailed-targeting-info/detaile
 import { DetailedTargetingModeService } from '../detailed-targeting-mode/detailed-targeting-mode.service';
 import { DetailedTargetingDropdownBrowseService } from '../detailed-targeting-dropdown-browse/detailed-targeting-dropdown-browse.service';
 import { FB } from '../../fb/fb.interface';
+import { Subject } from 'rxjs';
+import { DetailedTargetingItem } from '../detailed-targeting-item';
 
 @Injectable()
 export class DetailedTargetingApiService {
@@ -18,26 +20,26 @@ export class DetailedTargetingApiService {
                private DetailedTargetingDropdownBrowseService: DetailedTargetingDropdownBrowseService,
                private DetailedTargetingModeService: DetailedTargetingModeService) {}
 
-  search (q: string, adaccountId = 'act_944874195534529') {
+  public search (q: string, adaccountId = 'act_944874195534529') {
     this.api.subscribe((FB: FB) => {
       FB.api(`/${adaccountId}/targetingsearch`, { q: q }, (response) => {
         this.DetailedTargetingDropdownSuggestedService.updateDropdown(response.data);
       });
     });
-  }
+  };
 
-  browse (adaccountId = 'act_944874195534529') {
+  public browse (adaccountId = 'act_944874195534529') {
     this.api.subscribe((FB: FB) => {
       FB.api(`/${adaccountId}/targetingbrowse`, {
         include_headers: false,
-        include_nodes: true
+        include_nodes:   true
       }, (response) => {
         this.DetailedTargetingDropdownBrowseService.updateDropdown(response.data);
       });
     });
-  }
+  };
 
-  suggest (targetingList: Array<Object> = [], adaccountId = 'act_944874195534529') {
+  public suggest (targetingList: Array<Object> = [], adaccountId = 'act_944874195534529') {
     this.api.subscribe((FB: FB) => {
       FB.api(`/${adaccountId}/targetingsuggestions`, {
         targeting_list: targetingList
@@ -45,6 +47,20 @@ export class DetailedTargetingApiService {
         this.DetailedTargetingDropdownSuggestedService.updateDropdown(response.data);
       });
     });
-  }
+  };
+
+  public validate (targetingList: Array<Object> = [], adaccountId = 'act_944874195534529') {
+    let _response = new Subject<DetailedTargetingItem[]>();
+
+    this.api.subscribe((FB: FB) => {
+      FB.api(`/${adaccountId}/targetingvalidation`, {
+        targeting_list: targetingList
+      }, (response) => {
+        _response.next(response.data);
+      });
+    });
+
+    return _response.asObservable();
+  };
 
 }