+++ src/app/detailed-targeting/detailed-targeting-selected/detailed-targeting-selected.component.ts
@@ -1,23 +1,22 @@
 import { Component, OnInit } from '@angular/core';
-import { TargetingSpecService, TargetingSpec } from '../../targeting/targeting-spec.service';
 import { DetailedTargetingSelectedService } from './detailed-targeting-selected.service';
 import { DetailedTargetingItem } from '../detailed-targeting-item';
 import { DetailedTargetingModeService } from '../detailed-targeting-mode/detailed-targeting-mode.service';
 import { DetailedTargetingDropdownBrowseService } from '../detailed-targeting-dropdown-browse/detailed-targeting-dropdown-browse.service';
+import { DetailedTargetingService } from '../detailed-targeting.service';
 
 @Component({
-  selector: 'detailed-targeting-selected',
+  selector:    'detailed-targeting-selected',
   templateUrl: 'detailed-targeting-selected.component.html',
-  styleUrls: ['detailed-targeting-selected.component.css']
+  styleUrls:   ['detailed-targeting-selected.component.css']
 })
 export class DetailedTargetingSelectedComponent implements OnInit {
 
-  private spec: TargetingSpec;
   private items: DetailedTargetingItem[];
 
   private structuredSelectedItems;
 
-  constructor (private TargetingSpecService: TargetingSpecService,
+  constructor (private DetailedTargetingService: DetailedTargetingService,
                private DetailedTargetingDropdownBrowseService: DetailedTargetingDropdownBrowseService,
                private DetailedTargetingModeService: DetailedTargetingModeService,
                private DetailedTargetingSelectedService: DetailedTargetingSelectedService) {
@@ -29,16 +28,17 @@ export class DetailedTargetingSelectedComponent implements OnInit {
    * @param index
    */
   public showCrumb (key: string, index: number) {
-    let path = key.split(' > ');
+    let path             = key.split(' > ');
     let defaultOpenItems = this.DetailedTargetingDropdownBrowseService.defaultOpenItems;
-    let openItems = (<any>Object).assign({}, defaultOpenItems);
+    // noinspection TypeScriptUnresolvedFunction
+    let openItems        = Object.assign({}, defaultOpenItems);
 
     path.forEach((crumb: string, pos: number) => {
       if (pos <= index) {
         let openItemKey = path.slice(0, pos + 1)
           .join(' > ');
 
-        openItems._scrollTo = openItemKey;
+        openItems._scrollTo    = openItemKey;
         openItems[openItemKey] = true;
       }
     });
@@ -49,7 +49,7 @@ export class DetailedTargetingSelectedComponent implements OnInit {
 
   public removeGroup (key) {
     let selectedItems = this.DetailedTargetingSelectedService.get();
-    let idsToRemove = this.structuredSelectedItems.map[key].map(item => item.id);
+    let idsToRemove   = this.structuredSelectedItems.map[key].map(item => item.id);
 
     selectedItems = selectedItems.filter(item => idsToRemove.indexOf(item.id) === -1);
 
@@ -65,15 +65,11 @@ export class DetailedTargetingSelectedComponent implements OnInit {
   }
 
   ngOnInit () {
-    this.TargetingSpecService.spec.subscribe((spec: TargetingSpec) => {
-      this.spec = spec;
-      console.log('Targeting Spec: ', this.spec);
-    });
-
-    this.DetailedTargetingSelectedService.items.subscribe((items: DetailedTargetingItem[]) => {
-      this.items = items;
-      this.TargetingSpecService.updateWithDetailedTargeting(this.items);
-    });
+    this.DetailedTargetingSelectedService.items
+      .subscribe((items: DetailedTargetingItem[]) => {
+        this.items = items;
+        this.DetailedTargetingService.updateWithSelectedItems(this.items);
+      });
 
     this.DetailedTargetingSelectedService.items
       .map(this.DetailedTargetingSelectedService.structureSelectedItems)