+++ src/app/targeting/targeting-spec.interface.ts
@@ -1,0 +1,4 @@
+import { DetailedTargetingSpec } from './targeting-spec-detailed.interface';
+export interface TargetingSpec extends DetailedTargetingSpec {
+  gender?: any;
+}