+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+<a name="0.0.6-10"></a>
+## [0.0.6-10](https://github.com/aitarget/aitarget-components/compare/v0.0.6-9...0.0.6-10) (2016-09-02)
+
+
+### Bug Fixes
+
+* **placeholder:** make input font-weight: normal, close #38 ([8ad7333](https://github.com/aitarget/aitarget-components/commit/8ad7333)), closes [#38](https://github.com/aitarget/aitarget-components/issues/38)
+
+
+### Features
+
+* **onchange:** trigger spec onchange function if it is really changed and skip triggering on initialization ([47cb9c6](https://github.com/aitarget/aitarget-components/commit/47cb9c6))
+* **spec:** when updating spec with new detailed targeting, don't leave properties with empty arrays ([0f0670e](https://github.com/aitarget/aitarget-components/commit/0f0670e))
+
+
+
+<a name="0.0.6-9"></a>
+## [0.0.6-9](https://github.com/aitarget/aitarget-components/compare/v0.0.6-9...0.0.6-9) (2016-08-31)
+
+
+### Bug Fixes
+
+* **placeholder:** make input font-weight: normal, close #38 ([8ad7333](https://github.com/aitarget/aitarget-components/commit/8ad7333)), closes [#38](https://github.com/aitarget/aitarget-components/issues/38)
+
+
+
 <a name="0.0.6-8"></a>
 ## [0.0.6-8](https://github.com/aitarget/aitarget-components/compare/v0.0.6-8...0.0.6-8) (2016-08-31)